fix file structure; add session package and set up server.js
