+++ main/controllers/homeRoutes.js
@@ -1,0 +1,7 @@
+const router = require("express").Router();
+
+router.get("/", async (req, res) => {
+  res.render("homepage");
+});
+
+module.exports = router;
