+++ main/controllers/index.js
@@ -1,6 +1,6 @@
 const router = require("express").Router();
-// const homeRoutes = require('./homeRoutes');
+const homeRoutes = require("./homeRoutes");
 
-// router.use('/', homeRoutes);
+router.use("/", homeRoutes);
 
 module.exports = router;
