+++ main/package.json
@@ -5,7 +5,8 @@
   "main": "server.js",
   "scripts": {
     "seed": "node db/seed.js",
-    "test": "test"
+    "test": "test",
+    "dev": "nodemon server.js"
   },
   "repository": {
     "type": "git",
@@ -15,9 +16,11 @@
   "license": "ISC",
   "dependencies": {
     "bcrypt": "^5.1.0",
+    "connect-session-sequelize": "^7.1.7",
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "express-handlebars": "^7.1.2",
+    "express-session": "^1.17.3",
     "mysql2": "^3.6.0",
     "sequelize": "^6.32.1"
   }
